Move Preferences Model into Common. Add Argument Validation to GetByIds Method. Improve Recommended Properties Query.
