+++ src/Properties/Properties.Api/Controllers/PropertiesController.cs
@@ -192,7 +192,7 @@ namespace BuildingMarket.Properties.Api.Controllers
             _logger.LogInformation($"Getting top 6 recommended properties.");
 
             var userId = User.Claims.First(x => x.Type == ClaimTypes.Sid).Value;
-            var properties = await _mediator.Send(new GetRecommendedQuery() { BuyerId = userId });
+            var properties = await _mediator.Send(new GetRecommendedQuery { BuyerId = userId });
 
             return Ok(properties);
         }
