+++ src/Properties/Properties.Application/Contracts/IRecommendationRepository.cs
@@ -1,9 +1,9 @@
-﻿using BuildingMarket.Properties.Application.Models.Security;
+﻿using BuildingMarket.Common.Models;
 
 namespace BuildingMarket.Properties.Application.Contracts
 {
     public interface IRecommendationRepository
     {
-        Task<IEnumerable<int>> GetRecommended(PreferencesModel preferences, CancellationToken cancellationToken);
+        Task<IEnumerable<int>> GetRecommended(BuyerPreferencesRedisModel preferences, CancellationToken cancellationToken);
     }
 }
